Fixed bugs related to not removing event listeners on the socket.
The way we converte socket events to Observable, the observable can be unsubscribed but the event listener still gets called. Nothing removed it. So we had some duplicate logic occasionally.
Using rxjs fromEvent fixes that because it removes the event listener on unsubscribe/complete.
Have to convert the socket into a compatible object first.
Converted to using ngUnsubscribe$ to unsubscribe all listeners (partly because I needed yet another subscription).